move connectsettings to new file
